根据数据库字段类型的修改，将对应类的相应字段进行了修改； 完成了高级查询服务类，还没进行调试；
git-svn-id: https://zuow-pc:8443/svn/navy@37 47ee70cb-904c-3e43-b3d1-3900189539e7

+++ navy_v3/src/cn/com/hd/dto/navy/TTransport.java
@@ -14,7 +14,7 @@ public class TTransport extends BaseDTO {
 	private String comid;
 	private String comname;
 	private String trucktype;
-	private Integer deadweight;
+	private Double deadweight;
 	private Integer count;
 	private String nearrailway;
 	private Double rwdis;
@@ -47,11 +47,11 @@ public class TTransport extends BaseDTO {
 		this.trucktype = trucktype;
 	}
 
-	public Integer getDeadweight() {
+	public Double getDeadweight() {
 		return deadweight;
 	}
 
-	public void setDeadweight(Integer deadweight) {
+	public void setDeadweight(Double deadweight) {
 		this.deadweight = deadweight;
 	}
